requirements.txt
fastapi
uvicorn
PyPDF2
sentence-transformers
numpy
python-dotenv
requests
rich